Move mediator tests to lib/
